+++ src/main/java/kitchenpos/order/delivery/application/DeliveryOrderService.java
@@ -1,0 +1,5 @@
+package kitchenpos.order.delivery.application;
+
+public class DeliveryOrderService {
+    // Not yet implemented
+}
