+++ src/main/java/kitchenpos/order/delivery/infra/DefaultKitchenridersClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.infra;
+package kitchenpos.order.delivery.infra;
 
 import org.springframework.stereotype.Component;
