+++ src/main/java/kitchenpos/order/eatin/application/EatInOrderService.java
@@ -1,0 +1,5 @@
+package kitchenpos.order.eatin.application;
+
+public class EatInOrderService {
+    // Not yet implemented
+}
