+++ src/main/java/kitchenpos/order/eatin/domain/OrderTableRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.domain;
+package kitchenpos.order.eatin.domain;
 
 import java.util.List;
 import java.util.Optional;
