+++ src/main/java/kitchenpos/order/order/domain/JpaOrderRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.domain;
+package kitchenpos.order.order.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
