+++ src/main/java/kitchenpos/order/order/domain/Order.java
@@ -1,4 +1,6 @@
-package kitchenpos.order.domain;
+package kitchenpos.order.order.domain;
+
+import kitchenpos.order.eatin.domain.OrderTable;
 
 import javax.persistence.*;
 import java.time.LocalDateTime;
