+++ src/main/java/kitchenpos/order/order/domain/OrderType.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.domain;
+package kitchenpos.order.order.domain;
 
 public enum OrderType {
     DELIVERY, TAKEOUT, EAT_IN
