+++ src/main/java/kitchenpos/order/takeout/application/TakeoutOrderService.java
@@ -1,0 +1,5 @@
+package kitchenpos.order.takeout.application;
+
+public class TakeoutOrderService {
+    // Not yet implemented
+}
